ref: Moved all misc custom functions to its own file
